server/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyMySQL
passlib[bcrypt]
PyJWT
python-socketio
python-dotenv
pydantic
